requirements.txt
asgiref~=3.9
certifi~=2025.8
charset-normalizer~=3.4
crispy-bootstrap5~=2024.2
defusedxml~=0.7
Django~=5.2
django-admin-index~=3.1
django-admin-interface~=0.28
django-bootstrap-icons~=0.8
django-colorfield~=0.11
django-crispy-forms~=2.1
django-filter~=24.1
django-generic-positions~=0.2
django-hvad~=1.8
django-jazzmin~=2.6
django-libs~=2.0
django-ordered-model~=3.7
django-review~=1.10
django-user-media~=1.2
easy-thumbnails~=2.8
idna~=3.8
pillow~=11.3
python-slugify~=8.0
requests~=2.32
simplejson~=3.19
sqlparse~=0.5
text-unidecode~=1.3
tzdata~=2025.2
urllib3~=2.2
